refactor(iori): replace dash live with live2
Thank you, Jules!
