+++ crates/iori/src/dash/mod.rs
@@ -1,6 +1,5 @@
 pub mod archive;
 pub mod live;
-pub mod live2;
 pub mod segment;
 pub mod template;
 pub(crate) mod url;
